+++ lib/Auth/Process/DatabaseCommand.php
@@ -13,6 +13,7 @@ class DatabaseCommand
 {
     private $databaseConnector;
     private $conn;
+    private $dbDriver;
     private $statisticsTableName;
     private $detailedStatisticsTableName;
     private $identityProvidersMapTableName;
@@ -23,6 +24,7 @@ class DatabaseCommand
         $this->databaseConnector = new DatabaseConnector();
         $this->conn = $this->databaseConnector->getConnection();
         assert($this->conn !== null);
+        $this->dbDriver = $this->databaseConnector->getDbDriver();
         $this->statisticsTableName = $this->databaseConnector->getStatisticsTableName();
         $this->detailedStatisticsTableName = $this->databaseConnector->getDetailedStatisticsTableName();
         $this->identityProvidersMapTableName = $this->databaseConnector->getIdentityProvidersMapTableName();
@@ -43,7 +45,12 @@ class DatabaseCommand
         if ($user && $this->databaseConnector->getDetailedDays() > 0) {
             // write also into aggregated statistics
             self::writeLogin($year, $month, $day, $sourceIdp, $service);
-            $params['user'] = $user;
+            if ($this->dbDriver == 'pgsql') {
+                // the word "user" is reserved from PostgreSQL
+                $params['userid'] = $user;
+            } else {
+                $params['user'] = $user;
+            }
             $table = $this->detailedStatisticsTableName;
         }
         $fields = array_keys($params);
@@ -51,8 +58,20 @@ class DatabaseCommand
             return ':' . $field;
 
         }, $fields);
-        $query = "INSERT INTO " . $table . " (" . implode(', ', $fields) . ")" .
-                 " VALUES (" . implode(', ', $placeholders) . ") ON DUPLICATE KEY UPDATE count = count + 1";
+        if ($this->dbDriver == 'pgsql') {
+            // remove count field from ON CONFLICT statement
+            $conflictFields = $fields;
+            $pos = array_search('count', $conflictFields);
+            unset($conflictFields[$pos]);
+            $query = "INSERT INTO " . $table . " (" . implode(', ', $fields) . ")" .
+                     " VALUES (" . implode(', ', $placeholders) . ")" .
+                     " ON CONFLICT (" . implode(', ', $conflictFields) . ")" .
+                     " DO UPDATE SET count =  " . $table . ".count + 1";
+        } else {
+            $query = "INSERT INTO " . $table . " (" . implode(', ', $fields) . ")" .
+                     " VALUES (" . implode(', ', $placeholders) . ")" .
+                     " ON DUPLICATE KEY UPDATE count = count + 1";
+        }
 
         return $this->conn->write($query, $params);
     }
@@ -96,17 +115,29 @@ class DatabaseCommand
             }
 
             if (!empty($idpName)) {
+                if ($this->dbDriver == 'pgsql') {
+                    $query = "INSERT INTO " . $this->identityProvidersMapTableName .
+                             " (entityId, name) VALUES (:idp, :name1) ON CONFLICT (entityId) DO UPDATE SET name = :name2";
+                } else {
+                    $query = "INSERT INTO " . $this->identityProvidersMapTableName .
+                             " (entityId, name) VALUES (:idp, :name1) ON DUPLICATE KEY UPDATE name = :name2";
+                }
                 $this->conn->write(
-                    "INSERT INTO " . $this->identityProvidersMapTableName .
-                    "(entityId, name) VALUES (:idp, :name1) ON DUPLICATE KEY UPDATE name = :name2",
+                    $query,
                     ['idp'=>$idpEntityID, 'name1'=>$idpName, 'name2'=>$idpName]
                 );
             }
 
             if (!empty($spName)) {
+                if ($this->dbDriver == 'pgsql') {
+                    $query = "INSERT INTO " . $this->serviceProvidersMapTableName .
+                             " (identifier, name) VALUES (:sp, :name1) ON CONFLICT (identifier) DO UPDATE SET name = :name2";
+                } else {
+                    $query = "INSERT INTO " . $this->serviceProvidersMapTableName .
+                             " (identifier, name) VALUES (:sp, :name1) ON DUPLICATE KEY UPDATE name = :name2";
+                }
                 $this->conn->write(
-                    "INSERT INTO " . $this->serviceProvidersMapTableName .
-                    "(identifier, name) VALUES (:sp, :name1) ON DUPLICATE KEY UPDATE name = :name2",
+                    $query,
                     ['sp'=>$spEntityId, 'name1'=>$spName, 'name2'=>$spName]
                 );
             }
@@ -140,7 +171,7 @@ class DatabaseCommand
                  "FROM " . $this->statisticsTableName . " " .
                  "WHERE service != '' ";
         $params = [];
-        self::addDaysRange($days, $query, $params);
+        self::addDaysRange($days, $this->dbDriver, $query, $params);
         $query .= "GROUP BY year,month,day " .
                   "ORDER BY year ASC,month ASC,day ASC";
 
@@ -153,7 +184,7 @@ class DatabaseCommand
                  "FROM " . $this->statisticsTableName . " " .
                  "WHERE service=:service ";
         $params = ['service' => $spIdentifier];
-        self::addDaysRange($days, $query, $params);
+        self::addDaysRange($days, $this->dbDriver, $query, $params);
         $query .= "GROUP BY year,month,day " .
                   "ORDER BY year ASC,month ASC,day ASC";
 
@@ -166,7 +197,7 @@ class DatabaseCommand
                  "FROM " . $this->statisticsTableName . " " .
                  "WHERE sourceIdP=:sourceIdP ";
         $params = ['sourceIdP'=>$idpIdentifier];
-        self::addDaysRange($days, $query, $params);
+        self::addDaysRange($days, $this->dbDriver, $query, $params);
         $query .= "GROUP BY year,month,day " .
                   "ORDER BY year ASC,month ASC,day ASC";
 
@@ -175,57 +206,93 @@ class DatabaseCommand
 
     public function getAccessCountPerService($days)
     {
-        $query = "SELECT IFNULL(name,service) AS spName, service, SUM(count) AS count " .
-                 "FROM " . $this->serviceProvidersMapTableName . " " .
-                 "LEFT OUTER JOIN " . $this->statisticsTableName . " ON service = identifier ";
+        if ($this->dbDriver == 'pgsql') {
+            $query = "SELECT COALESCE(name,service) AS spname, service, SUM(count) AS count " .
+                     "FROM " . $this->serviceProvidersMapTableName . " " .
+                     "LEFT OUTER JOIN " . $this->statisticsTableName . " ON service = identifier ";
+            $querySuffix = "GROUP BY service, name HAVING service != '' " .
+                           "ORDER BY count DESC";
+        } else {
+            $query = "SELECT IFNULL(name,service) AS spName, service, SUM(count) AS count " .
+                     "FROM " . $this->serviceProvidersMapTableName . " " .
+                     "LEFT OUTER JOIN " . $this->statisticsTableName . " ON service = identifier ";
+            $querySuffix = "GROUP BY service HAVING service != '' " .
+                           "ORDER BY count DESC";
+        }
         $params = [];
-        self::addDaysRange($days, $query, $params);
-        $query .= "GROUP BY service HAVING service != '' " .
-                  "ORDER BY count DESC";
+        self::addDaysRange($days, $this->dbDriver, $query, $params);
+        $query .= $querySuffix;
 
         return $this->conn->read($query, $params)->fetchAll(PDO::FETCH_NUM);
     }
 
     public function getAccessCountForServicePerIdentityProviders($days, $spIdentifier)
     {
-        $query = "SELECT IFNULL(name,sourceIdp) AS idpName, SUM(count) AS count " .
-                 "FROM " . $this->identityProvidersMapTableName . " " .
-                 "LEFT OUTER JOIN " . $this->statisticsTableName . " ON sourceIdp = entityId ";
+        if ($this->dbDriver == 'pgsql') {
+            $query = "SELECT COALESCE(name,sourceIdp) AS idpname, SUM(count) AS count " .
+                     "FROM " . $this->identityProvidersMapTableName . " " .
+                     "LEFT OUTER JOIN " . $this->statisticsTableName . " ON sourceIdp = entityId ";
+            $querySuffix = "GROUP BY sourceIdp, service, idpname HAVING sourceIdp != '' AND service=:service " .
+                           "ORDER BY count DESC";
+        } else {
+            $query = "SELECT IFNULL(name,sourceIdp) AS idpname, SUM(count) AS count " .
+                     "FROM " . $this->identityProvidersMapTableName . " " .
+                     "LEFT OUTER JOIN " . $this->statisticsTableName . " ON sourceIdp = entityId ";
+            $querySuffix = "GROUP BY sourceIdp, service HAVING sourceIdp != '' AND service=:service " .
+                           "ORDER BY count DESC";
+        }
         $params = ['service' => $spIdentifier];
-        self::addDaysRange($days, $query, $params);
-        $query .= "GROUP BY sourceIdp, service HAVING sourceIdp != '' AND service=:service " .
-                  "ORDER BY count DESC";
+        self::addDaysRange($days, $this->dbDriver, $query, $params);
+        $query .= $querySuffix;
 
         return $this->conn->read($query, $params)->fetchAll(PDO::FETCH_NUM);
     }
 
     public function getAccessCountForIdentityProviderPerServiceProviders($days, $idpEntityId)
     {
-        $query = "SELECT IFNULL(name,service) AS spName, SUM(count) AS count " .
-                 "FROM " . $this->serviceProvidersMapTableName . " " .
-                 "LEFT OUTER JOIN " . $this->statisticsTableName . " ON service = identifier ";
+        if ($this->dbDriver == 'pgsql') {
+            $query = "SELECT COALESCE(name,service) AS spname, SUM(count) AS count " .
+                     "FROM " . $this->serviceProvidersMapTableName . " " .
+                     "LEFT OUTER JOIN " . $this->statisticsTableName . " ON service = identifier ";
+            $querySuffix = "GROUP BY sourceIdp, service, name HAVING service != '' AND sourceIdp=:sourceIdp " .
+                           "ORDER BY count DESC";
+        } else {
+            $query = "SELECT IFNULL(name,service) AS spname, SUM(count) AS count " .
+                     "FROM " . $this->serviceProvidersMapTableName . " " .
+                     "LEFT OUTER JOIN " . $this->statisticsTableName . " ON service = identifier ";
+            $querySuffix = "GROUP BY sourceIdp, service HAVING service != '' AND sourceIdp=:sourceIdp " .
+                           "ORDER BY count DESC";
+        }
         $params = ['sourceIdp'=>$idpEntityId];
-        self::addDaysRange($days, $query, $params);
-        $query .= "GROUP BY sourceIdp, service HAVING service != '' AND sourceIdp=:sourceIdp " .
-                  "ORDER BY count DESC";
+        self::addDaysRange($days, $this->dbDriver, $query, $params);
+        $query .= $querySuffix;
 
         return $this->conn->read($query, $params)->fetchAll(PDO::FETCH_NUM);
     }
 
     public function getLoginCountPerIdp($days)
     {
-        $query = "SELECT IFNULL(name,sourceIdp) AS idpName, sourceIdp, SUM(count) AS count " .
-                 "FROM " . $this->identityProvidersMapTableName . " " .
-                 "LEFT OUTER JOIN " . $this->statisticsTableName . " ON sourceIdp = entityId ";
+        if ($this->dbDriver == 'pgsql') {
+            $query = "SELECT COALESCE(name,sourceIdp) AS idpname, sourceidp, SUM(count) AS count " .
+                     "FROM " . $this->identityProvidersMapTableName . " " .
+                     "LEFT OUTER JOIN " . $this->statisticsTableName . " ON sourceIdp = entityId ";
+            $querySuffix = "GROUP BY sourceidp, name HAVING sourceIdp != '' " .
+                           "ORDER BY count DESC";
+        } else {
+            $query = "SELECT IFNULL(name,sourceIdp) AS idpName, sourceIdp, SUM(count) AS count " .
+                     "FROM " . $this->identityProvidersMapTableName . " " .
+                     "LEFT OUTER JOIN " . $this->statisticsTableName . " ON sourceIdp = entityId ";
+            $querySuffix = "GROUP BY sourceIdp HAVING sourceIdp != '' " .
+                           "ORDER BY count DESC";
+        }
         $params = [];
-        self::addDaysRange($days, $query, $params);
-        $query .= "GROUP BY sourceIdp HAVING sourceIdp != '' " .
-                  "ORDER BY count DESC";
+        self::addDaysRange($days, $this->dbDriver, $query, $params);
+        $query .= $querySuffix;
 
         return $this->conn->read($query, $params)->fetchAll(PDO::FETCH_NUM);
     }
 
-    private static function addDaysRange($days, &$query, &$params, $not = false)
+    private static function addDaysRange($days, $dbDriver, &$query, &$params, $not = false)
     {
         if ($days != 0) {    // 0 = all time
             if (stripos($query, "WHERE") === false) {
@@ -233,11 +300,17 @@ class DatabaseCommand
             } else {
                 $query .= "AND";
             }
-            $query .= " CONCAT(year,'-',LPAD(month,2,'00'),'-',LPAD(day,2,'00')) ";
+            if ($dbDriver == 'pgsql') {
+                $query .= " CAST(CONCAT(year,'-',LPAD(CAST(month AS varchar),2,'0'),'-',LPAD(CAST(day AS varchar),2,'0')) AS date) ";
+                $querySuffix = "> current_date - INTERVAL '1 days' * :days ";
+            } else {
+                $query .= " CONCAT(year,'-',LPAD(month,2,'00'),'-',LPAD(day,2,'00')) ";
+                $querySuffix = "BETWEEN CURDATE() - INTERVAL :days DAY AND CURDATE() ";
+            }
             if ($not) {
                 $query .= "NOT ";
             }
-            $query .= "BETWEEN CURDATE() - INTERVAL :days DAY AND CURDATE() ";
+            $query .= $querySuffix;
             $params['days'] = $days;
         }
     }
@@ -247,7 +320,7 @@ class DatabaseCommand
         if ($this->databaseConnector->getDetailedDays() > 0) {
             $query = "DELETE FROM " . $this->detailedStatisticsTableName . " ";
             $params = [];
-            self::addDaysRange($this->databaseConnector->getDetailedDays(), $query, $params, true);
+            self::addDaysRange($this->databaseConnector->getDetailedDays(), $this->dbDriver, $query, $params, true);
             return $this->conn->write($query, $params);
         }
     }